requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
dj-database-url==1.2.0
Django==3.2.15
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-environ==0.10.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2023.3
schedule==1.2.0
six==1.16.0
sqlparse==0.4.4
whitenoise==6.3.0
